requirements-all.txt
-r requirements.txt
bitsandbytes==0.41.0    # quantization
scipy                   # required by bitsandbytes
sentencepiece           # llama-based models
tokenizers              # pythia, falcon, redpajama
datasets                # eval
zstandard               # scripts/prepare_redpajama.py, scripts/prepare_starcoder.py
pandas                  # scripts/prepare_csv.py, scripts/prepare_starcoder.py
pyarrow                 # scripts/prepare_starcoder.py
tensorboard             # pretrain/tinyllama.py
torchmetrics            # pretrain/tinyllama.py
# eval
git+https://github.com/EleutherAI/lm-evaluation-harness.git@115206dc89dad67b8beaa90051fb52db77f0a529
# scripts/prepare_slimpajama.py, scripts/prepare_starcoder.py, pretrain/tinyllama.py
lightning[data] @ git+https://github.com/Lightning-AI/lightning@6dfa5cca9de5c28548eef5582a53c483b0eda66a
